requirements.txt
fastapi~=0.115.7
uvicorn
python-dotenv~=1.0.1
python-multipart
numpy~=2.0.2
pillow
pytest
pytest-mock
tensorflow~=2.18.0
keras
prometheus-client~=0.21.1
pynvml~=12.0.0